requirements.txt
fastapi
uvicorn
psycopg2
pydantic
pydantic-settings
sqlalchemy
email-validator
passlib[bcrypt]
python-jose[cryptography]
python-multipart